validator - persistence - channel - api - `page` to `u32`

+++ validator/src/infrastructure/persistence/channel/api.rs
@@ -15,7 +15,7 @@ pub struct ApiChannelRepository {
 
 impl ChannelRepository for ApiChannelRepository {
     fn all(&self, identity: &str) -> RepositoryFuture<Vec<Channel>> {
-        let list_page_url = |page: u64| {
+        let list_page_url = |page: u32| {
             format!(
                 "http://localhost:8005/channel/list?validator={}&page={}",
                 identity, page
